Configures CORS origin from environment variable
Updates the CORS configuration to read the origin from the
`CORS_ORIGIN` environment variable. This allows for dynamic
configuration of the allowed origin(s) based on the environment.
It splits the value by comma to allow multiple origins. If the
environment variable is not set, CORS is enabled for all origins.